Consolidate cost stuff
